make: run golangci-lint container if not installed
Add a separate golangci-lint target to the Makefile which runs either
the locally installed golangci-lint binary - in case it is found and
matches the requested version - or runs it as a container otherwise.

As part of that, also explicitly require the latest golangci-lint
version v1.44.2.

The same approach is already used in the hubble and cilium-cli repos.

Suggested-by: Robin Hahling <robin.hahling@gw-computing.net>
Signed-off-by: Tobias Klauser <tobias@cilium.io>